# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hariyo Mart Nepal Marketplace API
  version: 2.0.0
  description: Location-based multi-tenant farmer marketplace API for discovery, farmer SaaS, products and multi-seller orders.
servers:
  - url: http://localhost:4000
paths:
...
  /api/tenants/mine:
    get:
      summary: Current seller tenant and farms
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Tenant workspace}
  /api/tenants:
    get:
      summary: Admin tenant directory
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Tenant collection}
  /api/tenants/{id}/verify:
    patch:
      summary: Admin verify or reject seller tenant
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses:
        '200': {description: Updated tenant}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}